refactor(audio-player): change filename
