+++ src/utils/locales/pt_BR.json
@@ -194,5 +194,6 @@
   "app.userList.menu.makePresenter.label": "Tornar apresentador",
   "mobileSdk.audioPlayer.modal.title": "Audio player",
   "mobileSdk.audioPlayer.modal.desc": "A funcionalidade possibilita o compartilhamento de áudios sincronizados durante a sessão.",
-  "mobileSdk.audioPlayer.modal.desc2": "O áudio é controlado pelo apresentador da sessão."
+  "mobileSdk.audioPlayer.modal.desc2": "O áudio é controlado pelo apresentador da sessão.",
+  "mobileSdk.audioPlayer.modal.loading": "Não há nenhum áudio ativo..."
 }
